+++ config/src/main/java/com/alibaba/nacos/config/server/service/repository/extrnal/ExternalCommonPersistServiceImpl.java
@@ -14,24 +14,26 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.config.server.service.repository.embedded;
+package com.alibaba.nacos.config.server.service.repository.extrnal;
 
-import com.alibaba.nacos.common.notify.NotifyCenter;
-import com.alibaba.nacos.config.server.configuration.ConditionOnEmbeddedStorage;
+import com.alibaba.nacos.config.server.configuration.ConditionOnExternalStorage;
 import com.alibaba.nacos.config.server.constant.Constants;
-import com.alibaba.nacos.config.server.exception.NacosConfigException;
 import com.alibaba.nacos.config.server.model.TenantInfo;
-import com.alibaba.nacos.config.server.model.event.DerbyImportEvent;
 import com.alibaba.nacos.config.server.service.datasource.DataSourceService;
 import com.alibaba.nacos.config.server.service.datasource.DynamicDataSource;
-import com.alibaba.nacos.config.server.service.repository.OtherPersistService;
-import com.alibaba.nacos.config.server.service.sql.EmbeddedStorageContextUtils;
+import com.alibaba.nacos.config.server.service.repository.CommonPersistService;
+import com.alibaba.nacos.config.server.utils.LogUtil;
 import com.alibaba.nacos.plugin.datasource.MapperManager;
 import com.alibaba.nacos.plugin.datasource.constants.TableConstant;
 import com.alibaba.nacos.plugin.datasource.mapper.TenantInfoMapper;
 import com.alibaba.nacos.sys.env.EnvUtil;
 import org.springframework.context.annotation.Conditional;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.CannotGetJdbcConnectionException;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.support.TransactionTemplate;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -41,90 +43,70 @@ import java.util.Objects;
 import static com.alibaba.nacos.config.server.service.repository.RowMapperManager.TENANT_INFO_ROW_MAPPER;
 
 /**
- * EmbeddedOtherPersistServiceImpl.
+ * ExternalOtherPersistServiceImpl.
  *
  * @author lixiaoshuang
  */
-@SuppressWarnings({"PMD.MethodReturnWrapperTypeRule", "checkstyle:linelength"})
-@Conditional(value = ConditionOnEmbeddedStorage.class)
-@Service("embeddedOtherPersistServiceImpl")
-public class EmbeddedOtherPersistServiceImpl implements OtherPersistService {
+@SuppressWarnings(value = {"PMD.MethodReturnWrapperTypeRule", "checkstyle:linelength"})
+@Conditional(value = ConditionOnExternalStorage.class)
+@Service("externalOtherPersistServiceImpl")
+public class ExternalCommonPersistServiceImpl implements CommonPersistService {
     
     private DataSourceService dataSourceService;
     
-    private final DatabaseOperate databaseOperate;
+    protected JdbcTemplate jt;
+    
+    protected TransactionTemplate tjt;
     
     private MapperManager mapperManager;
     
-    /**
-     * The constructor sets the dependency injection order.
-     *
-     * @param databaseOperate {@link EmbeddedStoragePersistServiceImpl}
-     */
-    public EmbeddedOtherPersistServiceImpl(DatabaseOperate databaseOperate) {
-        this.databaseOperate = databaseOperate;
+    public ExternalCommonPersistServiceImpl() {
         this.dataSourceService = DynamicDataSource.getInstance().getDataSource();
+        this.jt = dataSourceService.getJdbcTemplate();
+        this.tjt = dataSourceService.getTransactionTemplate();
         Boolean isDataSourceLogEnable = EnvUtil.getProperty(Constants.NACOS_PLUGIN_DATASOURCE_LOG, Boolean.class,
                 false);
         this.mapperManager = MapperManager.instance(isDataSourceLogEnable);
-        NotifyCenter.registerToSharePublisher(DerbyImportEvent.class);
     }
     
     @Override
     public void insertTenantInfoAtomic(String kp, String tenantId, String tenantName, String tenantDesc,
             String createResoure, final long time) {
-        
-        TenantInfoMapper tenantInfoMapper = mapperManager.findMapper(dataSourceService.getDataSourceType(),
-                TableConstant.TENANT_INFO);
-        final String sql = tenantInfoMapper.insert(
-                Arrays.asList("kp", "tenant_id", "tenant_name", "tenant_desc", "create_source", "gmt_create",
-                        "gmt_modified"));
-        final Object[] args = new Object[] {kp, tenantId, tenantName, tenantDesc, createResoure, time, time};
-        
-        EmbeddedStorageContextUtils.addSqlContext(sql, args);
-        
         try {
-            boolean result = databaseOperate.update(EmbeddedStorageContextUtils.getCurrentSqlContext());
-            if (!result) {
-                throw new NacosConfigException("Namespace creation failed");
-            }
-        } finally {
-            EmbeddedStorageContextUtils.cleanAllContext();
+            TenantInfoMapper tenantInfoMapper = mapperManager.findMapper(dataSourceService.getDataSourceType(),
+                    TableConstant.TENANT_INFO);
+            jt.update(tenantInfoMapper.insert(
+                    Arrays.asList("kp", "tenant_id", "tenant_name", "tenant_desc", "create_source", "gmt_create",
+                            "gmt_modified")), kp, tenantId, tenantName, tenantDesc, createResoure, time, time);
+        } catch (DataAccessException e) {
+            LogUtil.FATAL_LOG.error("[db-error] " + e, e);
+            throw e;
         }
     }
     
     @Override
     public void removeTenantInfoAtomic(final String kp, final String tenantId) {
-        TenantInfoMapper tenantInfoMapper = mapperManager.findMapper(dataSourceService.getDataSourceType(),
-                TableConstant.TENANT_INFO);
-        
-        EmbeddedStorageContextUtils.addSqlContext(tenantInfoMapper.delete(Arrays.asList("kp", "tenant_id")), kp,
-                tenantId);
         try {
-            databaseOperate.update(EmbeddedStorageContextUtils.getCurrentSqlContext());
-        } finally {
-            EmbeddedStorageContextUtils.cleanAllContext();
+            TenantInfoMapper tenantInfoMapper = mapperManager.findMapper(dataSourceService.getDataSourceType(),
+                    TableConstant.TENANT_INFO);
+            jt.update(tenantInfoMapper.delete(Arrays.asList("kp", "tenant_id")), kp, tenantId);
+        } catch (CannotGetJdbcConnectionException e) {
+            LogUtil.FATAL_LOG.error("[db-error] " + e, e);
+            throw e;
         }
     }
     
     @Override
     public void updateTenantNameAtomic(String kp, String tenantId, String tenantName, String tenantDesc) {
-        
-        TenantInfoMapper tenantInfoMapper = mapperManager.findMapper(dataSourceService.getDataSourceType(),
-                TableConstant.TENANT_INFO);
-        final String sql = tenantInfoMapper.update(Arrays.asList("tenant_name", "tenant_desc", "gmt_modified"),
-                Arrays.asList("kp", "tenant_id"));
-        final Object[] args = new Object[] {tenantName, tenantDesc, System.currentTimeMillis(), kp, tenantId};
-        
-        EmbeddedStorageContextUtils.addSqlContext(sql, args);
-        
         try {
-            boolean result = databaseOperate.update(EmbeddedStorageContextUtils.getCurrentSqlContext());
-            if (!result) {
-                throw new NacosConfigException("Namespace update failed");
-            }
-        } finally {
-            EmbeddedStorageContextUtils.cleanAllContext();
+            TenantInfoMapper tenantInfoMapper = mapperManager.findMapper(dataSourceService.getDataSourceType(),
+                    TableConstant.TENANT_INFO);
+            jt.update(tenantInfoMapper.update(Arrays.asList("tenant_name", "tenant_desc", "gmt_modified"),
+                            Arrays.asList("kp", "tenant_id")), tenantName, tenantDesc, System.currentTimeMillis(), kp,
+                    tenantId);
+        } catch (DataAccessException e) {
+            LogUtil.FATAL_LOG.error("[db-error] " + e, e);
+            throw e;
         }
     }
     
@@ -134,8 +116,17 @@ public class EmbeddedOtherPersistServiceImpl implements OtherPersistService {
                 TableConstant.TENANT_INFO);
         String sql = tenantInfoMapper.select(Arrays.asList("tenant_id", "tenant_name", "tenant_desc"),
                 Collections.singletonList("kp"));
-        return databaseOperate.queryMany(sql, new Object[] {kp}, TENANT_INFO_ROW_MAPPER);
-        
+        try {
+            return this.jt.query(sql, new Object[] {kp}, TENANT_INFO_ROW_MAPPER);
+        } catch (CannotGetJdbcConnectionException e) {
+            LogUtil.FATAL_LOG.error("[db-error] " + e, e);
+            throw e;
+        } catch (EmptyResultDataAccessException e) {
+            return Collections.emptyList();
+        } catch (Exception e) {
+            LogUtil.FATAL_LOG.error("[db-other-error]" + e.getMessage(), e);
+            throw new RuntimeException(e);
+        }
     }
     
     @Override
@@ -144,8 +135,17 @@ public class EmbeddedOtherPersistServiceImpl implements OtherPersistService {
                 TableConstant.TENANT_INFO);
         String sql = tenantInfoMapper.select(Arrays.asList("tenant_id", "tenant_name", "tenant_desc"),
                 Arrays.asList("kp", "tenant_id"));
-        return databaseOperate.queryOne(sql, new Object[] {kp, tenantId}, TENANT_INFO_ROW_MAPPER);
-        
+        try {
+            return jt.queryForObject(sql, new Object[] {kp, tenantId}, TENANT_INFO_ROW_MAPPER);
+        } catch (CannotGetJdbcConnectionException e) {
+            LogUtil.FATAL_LOG.error("[db-error] " + e, e);
+            throw e;
+        } catch (EmptyResultDataAccessException e) {
+            return null;
+        } catch (Exception e) {
+            LogUtil.FATAL_LOG.error("[db-other-error]" + e.getMessage(), e);
+            throw new RuntimeException(e);
+        }
     }
     
     @Override
@@ -161,9 +161,9 @@ public class EmbeddedOtherPersistServiceImpl implements OtherPersistService {
     
     @Override
     public boolean isExistTable(String tableName) {
-        String sql = String.format("SELECT 1 FROM %s FETCH FIRST ROW ONLY", tableName);
+        String sql = String.format("SELECT 1 FROM %s LIMIT 1", tableName);
         try {
-            databaseOperate.queryOne(sql, Integer.class);
+            jt.queryForObject(sql, Integer.class);
             return true;
         } catch (Throwable e) {
             return false;
@@ -178,10 +178,10 @@ public class EmbeddedOtherPersistServiceImpl implements OtherPersistService {
         TenantInfoMapper tenantInfoMapper = mapperManager.findMapper(dataSourceService.getDataSourceType(),
                 TableConstant.TENANT_INFO);
         String sql = tenantInfoMapper.count(Arrays.asList("tenant_id"));
-        Integer result = databaseOperate.queryOne(sql, new String[] {tenantId}, Integer.class);
+        Integer result = this.jt.queryForObject(sql, new String[] {tenantId}, Integer.class);
         if (result == null) {
             return 0;
         }
-        return result;
+        return result.intValue();
     }
 }
